requirements.txt
openai-whisper
speechrecognition
pyaudio